+++ module/Revitalizer.js
@@ -1,8 +1,7 @@
-import { runPIR } from "../src/pf2e-item-revitalizer.js";
+import { runPIR, waitForElementToBeRendered } from "../src/pf2e-item-revitalizer.js";
 import { id as SCRIPT_ID, title as SCRIPT_NAME } from "../module.json";
 
 export class Revitalizer {
-    toggled = false;
     dialog = undefined;
     HTMLtemplate = undefined;
 
@@ -15,34 +14,41 @@ export class Revitalizer {
         Loading actors, and looking at their respective data...
         <div class="loader"></div>
     </section>`;
-        
-        //await renderTemplate("../templates/dialog.html");
     }
 
-    async toggleElement() {
+    async renderPirContainerElement() {
         console.log(`${SCRIPT_ID} | Toggling display`);
 
-        /*if (this.toggled) {
-            this.toggled = false;
-            calendar.close();
-        } else {
-            calendar.toggled = true;*/
-            this.dialog = await new Dialog({
-                title: SCRIPT_NAME,
-                content: this.HTMLtemplate,
-                buttons: {
-                    ok: {
-                        icon: '<i class="fas fa-times"></i>',
-                        label: "Done",
-                    },
+        this.dialog = await new Dialog({
+            title: SCRIPT_NAME,
+            content: this.HTMLtemplate,
+            buttons: {
+                ok: {
+                    icon: '<i class="fas fa-times"></i>',
+                    label: "Done",
                 },
-            }).render(true);
-
-            await runPIR()
-        //}
+            },
+        }).render(true);
     }
 
-    async runCalculation() {
-        return await this.toggleElement();
+    async runCalculation(runForAll) {
+        let pirContainerElement = await waitForElementToBeRendered("pir-container");
+
+        // Get all available actors
+        let actors = canvas.tokens.placeables.filter((token) => token.actor).map((token) => token.actor);
+        if (!runForAll)
+            // Get only PCs
+            actors = actors.filter((token) => token.type == "character");
+
+        // TODO check token ownership
+
+        // TODO add checklist for GM to select from available tokens
+
+        return await runPIR(actors)
+    }
+
+    async run(runForAll) {
+        await this.renderPirContainerElement()
+        return await this.runCalculation(runForAll);
     }
 }